+++ MarineLang/SyntaxAnalysis/SyntaxAnalyzer.cs
@@ -1,4 +1,5 @@
 ﻿using MarineLang.Models;
+using MarineLang.Models.Asts;
 using MarineLang.Streams;
 using System.Collections.Generic;
 using System.Linq;
