+++ MarineLang/VirtualMachines/FuncScopeVariables.cs
@@ -1,4 +1,4 @@
-﻿using MarineLang.Models;
+﻿using MarineLang.Models.Asts;
 using System.Collections.Generic;
 using System.Linq;
