+++ MarineLang/VirtualMachines/HighLevelVirtualMachine.cs
@@ -1,5 +1,5 @@
 ﻿using MarineLang.BuildInObjects;
-using MarineLang.Models;
+using MarineLang.Models.Asts;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
